+++ matryoshka_tester/data/containers-2.json
@@ -1,0 +1,17 @@
+### THIS IS NOT VALID I KNOW. JUST SHOWING what we could evolve to.
+[
+    {
+        "name": "optionalname-openjdk-devel-16",
+        "type": "openjdk-devel",
+        "registry": "registry.suse.de",
+        "repo": "devel/bci/images",
+        "image": "bci/openjdk-devel",
+        "tag": "16"
+    },
+    {
+        "type": "openjdk-devel",
+        "repo": "devel/bci/images",
+        "image": "bci/openjdk-devel",
+        "tag": "16"
+    },
+]
