+++ setup.cfg
@@ -3,6 +3,7 @@ name = matryoshka-tester
 
 [options]
 packages = matryoshka_tester
+include_package_data = True
 
 [options.entry_points]
 console_scripts =
